refactor(python): remove HubClient and integrate hub methods directly into Client and AsyncClient
This commit eliminates the HubClient class, integrating its functionality directly into the Client and AsyncClient classes. The methods for managing agents and skills (pull, push, delete, check existence, and list) are now directly accessible, streamlining the API.

Additionally, the import statements have been cleaned up, and corresponding unit tests have been added to ensure the correctness of the new implementations. This change aligns the Python SDK with the structure of the Java SDK and enhances usability.
