+++ python/tests/unit_tests/test_hub_methods.py
@@ -1,6 +1,7 @@
-"""Test the HubClient and AsyncHubClient classes for Hub non-prompt repos."""
+"""Test the hub (agent/skill) methods on Client and AsyncClient."""
 
 import json
+from types import MethodType
 from unittest.mock import AsyncMock, MagicMock
 
 import pydantic
@@ -8,7 +9,8 @@ import pytest
 
 from langsmith import schemas as ls_schemas
 from langsmith import utils as ls_utils
-from langsmith.hub_client import AsyncHubClient, HubClient
+from langsmith.async_client import AsyncClient
+from langsmith.client import Client
 
 TOOLS_JSON = json.dumps(
     {
@@ -18,6 +20,27 @@ TOOLS_JSON = json.dumps(
 )
 
 
+_HUB_METHODS = [
+    "pull_agent",
+    "push_agent",
+    "pull_skill",
+    "push_skill",
+    "delete_agent",
+    "delete_skill",
+    "agent_exists",
+    "skill_exists",
+    "list_agents",
+    "list_skills",
+    "_pull_hub_directory",
+    "_push_hub_directory",
+    "_delete_hub_directory",
+    "_list_hub_repos",
+    "_hub_repo_exists",
+    "_create_hub_repo",
+    "_update_hub_repo_metadata",
+]
+
+
 def _mock_sync_client() -> MagicMock:
     client = MagicMock()
     client._host_url = "https://smith.langchain.com"
@@ -25,6 +48,9 @@ def _mock_sync_client() -> MagicMock:
     client._owner_conflict_error.return_value = ls_utils.LangSmithUserError(
         "owner mismatch"
     )
+    # Bind real Client hub methods so they execute with the mock as self.
+    for name in _HUB_METHODS:
+        setattr(client, name, MethodType(getattr(Client, name), client))
     return client
 
 
@@ -36,6 +62,8 @@ def _mock_async_client() -> MagicMock:
         return_value=ls_utils.LangSmithUserError("owner mismatch")
     )
     client._arequest_with_retries = AsyncMock()
+    for name in _HUB_METHODS:
+        setattr(client, name, MethodType(getattr(AsyncClient, name), client))
     return client
 
 
@@ -75,7 +103,7 @@ def test_agent_entry_exclude_none_strips_response_only_fields() -> None:
 
 
 def test_push_agent_rejects_short_parent_commit() -> None:
-    ctx = HubClient(_mock_sync_client())
+    ctx = _mock_sync_client()
     with pytest.raises(ls_utils.LangSmithUserError, match="8-64"):
         ctx.push_agent("-/repo", files={}, parent_commit="abc")
 
@@ -85,7 +113,7 @@ def test_push_agent_rejects_invalid_handle_on_create() -> None:
     client.request_with_retries.side_effect = ls_utils.LangSmithNotFoundError(
         "not found"
     )
-    ctx = HubClient(client)
+    ctx = client
     with pytest.raises(ls_utils.LangSmithUserError, match="Invalid repo_handle"):
         ctx.push_agent(
             "-/BadName",
@@ -105,7 +133,7 @@ def test_push_agent_rejects_invalid_handle_on_create() -> None:
     ],
 )
 def test_repo_handle_pattern(handle: str, valid: bool) -> None:
-    from langsmith.hub_client import _REPO_HANDLE_PATTERN
+    from langsmith.client import _REPO_HANDLE_PATTERN
 
     assert bool(_REPO_HANDLE_PATTERN.match(handle)) is valid
 
@@ -119,7 +147,7 @@ def test_pull_agent_hits_correct_url_and_parses() -> None:
             "files": {"main.py": {"type": "file", "content": "print('hi')"}},
         }
     )
-    ctx = HubClient(client)
+    ctx = client
     agent = ctx.pull_agent("owner/my-agent")
     call = client.request_with_retries.call_args
     assert call.args == ("GET", "/v1/platform/hub/repos/owner/my-agent/directories")
@@ -140,7 +168,7 @@ def test_pull_skill_returns_skill_context() -> None:
             "files": {},
         }
     )
-    ctx = HubClient(client)
+    ctx = client
     assert isinstance(ctx.pull_skill("o/s"), ls_schemas.SkillContext)
 
 
@@ -153,7 +181,7 @@ def test_pull_agent_with_version_passes_commit_param() -> None:
             "files": {},
         }
     )
-    ctx = HubClient(client)
+    ctx = client
     ctx.pull_agent("owner/repo", version="abc12345")
     assert client.request_with_retries.call_args.kwargs["params"] == {
         "commit": "abc12345"
@@ -174,7 +202,7 @@ def test_push_agent_creates_new_repo_and_commits() -> None:
             }
         ),
     ]
-    ctx = HubClient(client)
+    ctx = client
     url = ctx.push_agent(
         "-/my-agent", files={"main.py": ls_schemas.FileEntry(content="x")}
     )
@@ -210,7 +238,7 @@ def test_push_agent_updates_metadata_when_repo_exists() -> None:
             }
         ),
     ]
-    ctx = HubClient(client)
+    ctx = client
     ctx.push_agent(
         "-/my-agent",
         files={"main.py": ls_schemas.FileEntry(content="x")},
@@ -235,7 +263,7 @@ def test_push_agent_null_entry_serializes_as_delete() -> None:
             }
         ),
     ]
-    ctx = HubClient(client)
+    ctx = client
     ctx.push_agent(
         "-/my-agent",
         files={
@@ -262,7 +290,7 @@ def test_push_agent_accepts_tools_json_as_normal_file_entry() -> None:
             }
         ),
     ]
-    ctx = HubClient(client)
+    ctx = client
     ctx.push_agent(
         "-/my-agent",
         files={"tools.json": ls_schemas.FileEntry(content=TOOLS_JSON)},
@@ -286,7 +314,7 @@ def test_push_agent_passes_parent_commit_when_provided() -> None:
             }
         ),
     ]
-    ctx = HubClient(client)
+    ctx = client
     ctx.push_agent(
         "-/my-agent",
         files={"main.py": ls_schemas.FileEntry(content="x")},
@@ -299,7 +327,7 @@ def test_push_agent_passes_parent_commit_when_provided() -> None:
 def test_delete_agent_hits_directories_delete() -> None:
     client = _mock_sync_client()
     client.request_with_retries.return_value = _response({})
-    ctx = HubClient(client)
+    ctx = client
     ctx.delete_agent("-/old-agent")
     call = client.request_with_retries.call_args
     assert call.args == (
@@ -311,7 +339,7 @@ def test_delete_agent_hits_directories_delete() -> None:
 def test_agent_exists_returns_true_on_200() -> None:
     client = _mock_sync_client()
     client.request_with_retries.return_value = _response({})
-    ctx = HubClient(client)
+    ctx = client
     assert ctx.agent_exists("-/my-agent") is True
 
 
@@ -320,14 +348,14 @@ def test_agent_exists_returns_false_on_not_found() -> None:
     client.request_with_retries.side_effect = ls_utils.LangSmithNotFoundError(
         "not found"
     )
-    ctx = HubClient(client)
+    ctx = client
     assert ctx.agent_exists("-/nope") is False
 
 
 def test_skill_exists_returns_true_on_200() -> None:
     client = _mock_sync_client()
     client.request_with_retries.return_value = _response({})
-    ctx = HubClient(client)
+    ctx = client
     assert ctx.skill_exists("-/my-skill") is True
 
 
@@ -336,14 +364,14 @@ def test_skill_exists_returns_false_on_not_found() -> None:
     client.request_with_retries.side_effect = ls_utils.LangSmithNotFoundError(
         "not found"
     )
-    ctx = HubClient(client)
+    ctx = client
     assert ctx.skill_exists("-/nope") is False
 
 
 def test_list_agents_passes_repo_type_filter() -> None:
     client = _mock_sync_client()
     client.request_with_retries.return_value = _response({"repos": [], "total": 0})
-    ctx = HubClient(client)
+    ctx = client
     ctx.list_agents(limit=50, is_public=True, query="foo")
     call = client.request_with_retries.call_args
     assert call.args == ("GET", "/repos")
@@ -364,7 +392,7 @@ async def test_async_pull_agent_parses_and_merges_identifier() -> None:
             "files": {"main.py": {"type": "file", "content": "hi"}},
         }
     )
-    ctx = AsyncHubClient(client)
+    ctx = client
     agent = await ctx.pull_agent("owner/my-agent")
     assert isinstance(agent, ls_schemas.AgentContext)
     assert agent.owner == "owner"
@@ -390,7 +418,7 @@ async def test_async_push_agent_creates_and_commits() -> None:
             }
         ),
     ]
-    ctx = AsyncHubClient(client)
+    ctx = client
     url = await ctx.push_agent(
         "-/my-agent", files={"main.py": ls_schemas.FileEntry(content="x")}
     )
@@ -412,7 +440,7 @@ async def test_async_push_agent_accepts_tools_json_as_normal_file_entry() -> Non
             }
         ),
     ]
-    ctx = AsyncHubClient(client)
+    ctx = client
     await ctx.push_agent(
         "-/my-agent",
         files={"tools.json": ls_schemas.FileEntry(content=TOOLS_JSON)},
@@ -437,7 +465,7 @@ async def test_async_create_repo_swallows_conflict() -> None:
             }
         ),
     ]
-    ctx = AsyncHubClient(client)
+    ctx = client
     url = await ctx.push_agent(
         "-/my-agent", files={"main.py": ls_schemas.FileEntry(content="x")}
     )
@@ -448,7 +476,7 @@ async def test_async_create_repo_swallows_conflict() -> None:
 async def test_async_delete_agent_hits_directories_delete() -> None:
     client = _mock_async_client()
     client._arequest_with_retries.return_value = _response({})
-    ctx = AsyncHubClient(client)
+    ctx = client
     await ctx.delete_agent("-/old-agent")
     client._arequest_with_retries.assert_awaited_once_with(
         "DELETE",
@@ -459,7 +487,7 @@ async def test_async_delete_agent_hits_directories_delete() -> None:
 async def test_async_agent_exists_returns_true_on_200() -> None:
     client = _mock_async_client()
     client._arequest_with_retries.return_value = _response({})
-    ctx = AsyncHubClient(client)
+    ctx = client
     assert await ctx.agent_exists("-/my-agent") is True
 
 
@@ -468,14 +496,14 @@ async def test_async_agent_exists_returns_false_on_not_found() -> None:
     client._arequest_with_retries.side_effect = ls_utils.LangSmithNotFoundError(
         "not found"
     )
-    ctx = AsyncHubClient(client)
+    ctx = client
     assert await ctx.agent_exists("-/nope") is False
 
 
 async def test_async_skill_exists_returns_true_on_200() -> None:
     client = _mock_async_client()
     client._arequest_with_retries.return_value = _response({})
-    ctx = AsyncHubClient(client)
+    ctx = client
     assert await ctx.skill_exists("-/my-skill") is True
 
 
@@ -484,14 +512,14 @@ async def test_async_skill_exists_returns_false_on_not_found() -> None:
     client._arequest_with_retries.side_effect = ls_utils.LangSmithNotFoundError(
         "not found"
     )
-    ctx = AsyncHubClient(client)
+    ctx = client
     assert await ctx.skill_exists("-/nope") is False
 
 
 async def test_async_list_skills_passes_filter() -> None:
     client = _mock_async_client()
     client._arequest_with_retries.return_value = _response({"repos": [], "total": 0})
-    ctx = AsyncHubClient(client)
+    ctx = client
     await ctx.list_skills(limit=25)
     call = client._arequest_with_retries.call_args
     assert call.args == ("GET", "/repos")
